Added a master test script to run all tests

+++ front_end/Tests/logout/logout_accepted_when_login/test_etf.sh
@@ -2,7 +2,9 @@
 red='\033[0;31m'
 green='\033[0;32m'
 NC='\033[0m' # No Color
-rm transactions.etf
+if [  -f transactions.etf ]; then
+    rm transactions.etf
+fi
 ./../../../a.out <input.inp >output.bto
 DIFF=$(diff -q -w transactions.etf expected_transactions.etf)
 if [ "$DIFF" = "" ];then
